+++ tests/test_dns.py
@@ -15,6 +15,7 @@ EXPECTED_RECORDS = [
     "metrics.internal.lab.infiniteroomlabs.cloud",
     "alerts.internal.lab.infiniteroomlabs.cloud",
     "context.internal.lab.infiniteroomlabs.cloud",
+    "home.lab.infiniteroomlabs.cloud",
 ]